kerenl: sched: Make iowait_boost optional in schedutil
* We should apply the iowait boost only if cpufreq policy has iowait
boost enabled. Also make it a schedutil configuration from sysfs so it
can be turned on/off if needed (by default initialize it to the policy
value).

* For systems that don't need/want it enabled, such as those on arm64
based mobile devices that are battery operated, it saves energy when the
cpufreq driver policy doesn't have it enabled (details below):

* Here are some results for energy measurements collected running a
YouTube video for 30 seconds:
Before: 8.042533 mWh
After: 7.948377 mWh
Energy savings is ~1.2%

Bug: 38010527

Link: https://lkml.org/lkml/2017/5/19/42
Change-Id: If124076ad0c16ade369253840dedfbf870aff927
Signed-off-by: Joel Fernandes <joelaf@google.com>