kernel: sched: tune: Allow negative cpu boosts
* schedtune sets 0 as the floor for calculating CPU max boost, so
negative schedtune.boost values do not affect CPU frequency decisions.
Remove this restriction to allow userspace to apply negative boosts when
appropriate. Also change treatment of the root boost group to match the
other groups, so it only affects the CPU boost if it has runnable tasks
on the CPU.

* Test: set all boosts negative; sched_boost_cpu trace events show
negative CPU margins.

Change-Id: I89f3470299aef96a18797c105f02ebc8f367b5e1
Signed-off-by: Connor O'Brien <connoro@google.com>