kernel: sched: cpufreq_schedutil: Use >= when aggregating CPU loads in a policy
* When the normal utilization value for all the CPUs in a policy is 0,
the current aggregation scheme of using a > check will result in the
aggregated utilization being 0 and the max value being 1.

* This is not a problem for upstream code. However, since we also use
other statistics provided by WALT to update the aggregated utilization
value across all CPUs in a policy, we can end up with a non-zero
aggregated utilization while the max remains as 1.

* Then when get_next_freq() tries to compute the frequency using:
max-frequency * 1.25 * (util / max) it ends up with a frequency that is
greater than max-frequency. So the policy frequency jumps to
max-frequency.

* By changing the aggregation check to >=, we make sure that the max
value is updated with something reported by the scheduler for a CPU in
that policy. With the correct max, we can continue using the WALT
specific statistics without spurious jumps to max-frequency.

Change-Id: I14996cd796191192ea112f949dc42450782105f7
Signed-off-by: Saravana Kannan <skannan@codeaurora.org>