kernel: sched: Move schedtune en/dequeue before schedutil update triggers
* CPU rq util updates happen when rq signals are updated as part of
enqueue and dequeue operations. Doing these updates triggers a call to
the registered util update handler, which takes schedtune boosting into
account. Enqueueing the task in the correct schedtune group after this
happens means that we will potentially not see the boost for an entire
throttle period.

* Move the enqueue/dequeue operations for schedtune before the signal
updates which can trigger OPP changes.

Change-Id: I4236e6b194bc5daad32ff33067d4be1987996780
Signed-off-by: Chris Redpath <chris.redpath@arm.com>