kernel: sched: Add hold functionality to schedtune CPU boost
* Render - added back missing header

* When tasks come and go from a runqueue quickly, this can lead to boost
being applied and removed quickly which sometimes means we cannot raise
the CPU frequency again when we need to (due to the rate limit on
frequency updates). This has proved to be a particular issue for RT
tasks and alternative methods have been used in the past to work around
it.

* This is an attempt to solve the issue for all task classes and cpufreq
governors by introducing a generic mechanism in schedtune to retain the
max boost level from task enqueue for a minimum period - defined here as
50ms. This timeout was determined experimentally and is not
configurable.

* A sched_feat guards the application of this to tasks - in the default
configuration, task boosting only applied to tasks which have RT policy.
Change SCHEDTUNE_BOOST_HOLD_ALL to true to apply it to all tasks
regardless of class.

* It works like so:

* Every task enqueue (in an allowed class) stores a cpu-local timestamp.
If the task is not a member of an allowed class (all or RT depending
upon feature selection), the timestamp is not updated. The boost group
will stay active regardless of tasks present until 50ms beyond the last
timestamp stored. We also store the timestamp of the active boost group
to avoid unneccesarily revisiting the boost groups when checking CPU
boost level.

* If the timestamp is more than 50ms in the past when we check boost
then we re-evaluate the boost groups for that CPU, taking into account
the timestamps associated with each group.

* Idea based on rt-boost-retention patches from Joel.

Change-Id: I52cc2d2e82d1c5aa03550378c8836764f41630c1
Suggested-by: Joel Fernandes <joelaf@google.com>
Reviewed-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: Chris Redpath <chris.redpath@arm.com>
Signed-off-by: RenderBroken <zkennedy87@gmail.com>