kernel: sched: Use util_est in LB and WU paths
* joshuous: Adapted __cpu_util() to use delta instead of util, which is
in line with the change introduced in "sched: Extend sched_group_energy
to test load-balancing decisions" commit.

* When the scheduler looks at the CPU utilization, the current PELT
value for a CPU is returned straight away. In certain scenarios this can
have undesired side effects on task placement.

* For example, since the task utilization is decayed at wakeup time,
when a long sleeping big task is enqueued it does not add immediately a
significant contribution to the target CPU. As a result we generate a
race condition where other tasks can be placed on the same CPU while it
is still considered relatively empty.

* In order to reduce this kind of race conditions, this patch introduces
the required support to integrate the usage of the CPU's estimated
utilization in the wakeup path, via cpu_util_wake(), as well as in the
load-balance path, via cpu_util() which is used by update_sg_lb_stats().

* The estimated utilization of a CPU is defined to be the maximum
between its PELT's utilization and the sum of the estimated utilization
(at previous dequeue time) of all the tasks currently RUNNABLE on that
CPU. This allows to properly represent the spare capacity of a CPU
which, for example, has just got a big task running since a long sleep
period.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Reviewed-by: Dietmar Eggemann <dietmar.eggemann@arm.com>
Cc: Joel Fernandes <joelaf@google.com>
Cc: Juri Lelli <juri.lelli@redhat.com>
Cc: Linus Torvalds <torvalds@linux-foundation.org>
Cc: Morten Rasmussen <morten.rasmussen@arm.com>
Cc: Paul Turner <pjt@google.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Rafael J . Wysocki <rafael.j.wysocki@intel.com>
Cc: Steve Muckle <smuckle@google.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Todd Kjos <tkjos@android.com>
Cc: Vincent Guittot <vincent.guittot@linaro.org>
Cc: Viresh Kumar <viresh.kumar@linaro.org>
Link: http://lkml.kernel.org/r/20180309095245.11071-3-patrick.bellasi@arm.com
Signed-off-by: Ingo Molnar <mingo@kernel.org>
Signed-off-by: joshuous <joshuous@gmail.com>