kernel: sched: cpufreq_schedutil: Cache tunables on governor exit
* Currently when all the related CPUs from a policy go offline or the
governor is switched, cpufreq framework calls sugov_exit() that frees
the governor tunables. When any of the related CPUs comes back online or
governor is switched back to schedutil sugov_init() gets called which
allocates a fresh set of tunables that are set to default values. This
can cause the userspace settings to those tunables to be lost across
governor switches or when an entire cluster is hotplugged out. To
prevent this, save the tunable values on governor exit. Restore these
values to the newly allocated tunables on governor init.

Change-Id: I671d4d0e1a4e63e948bfddb0005367df33c0c249
Signed-off-by: Rohit Gupta <rohgup@codeaurora.org>
[Caching and restoring different tunables.]
Signed-off-by: joshuous <joshuous@gmail.com>